Naive Bayes with K means Clustering.
Graphical representations of the data show non-Gaussian distributions, which is a requirement for Naive Bayes. The unsupervised learning technique, k means, is used to decompose the classes into multiple Gaussian distributions, which assumes the data live on a Euclidean space and creates spherical clusters. The probability of belonging to each distribution is calculated and the data is categorised according to its highest probability.